+++ dev/bots/test/prepare_package_test.dart
@@ -10,8 +10,8 @@ import 'package:path/path.dart' as path;
 import 'package:platform/platform.dart' show FakePlatform, Platform;
 
 import '../../../packages/flutter_tools/test/src/fake_process_manager.dart';
-import '../prepare_package.dart';
 import '../prepare_package/archive_creator.dart';
+import '../prepare_package/archive_publisher.dart';
 import '../prepare_package/common.dart';
 import '../prepare_package/process_runner.dart';
 import 'common.dart';
